updated and merged
