+++ newshomepages/hyperlinks.py
@@ -2,13 +2,115 @@ import typing
 from pathlib import Path
 
 import click
-from bs4 import BeautifulSoup
 from playwright.sync_api import sync_playwright
 from playwright.sync_api._generated import BrowserContext
 from retry import retry
 from rich import print
 
 from . import utils
+
+get_link_divs_js = '''
+    var as = document.querySelectorAll('a')
+    as = Array.from(as)
+            .filter(function(a) { return a.href !== ''}).filter(function(a){return a.href !== undefined; })
+            .map(function(a) {return {'node': a, 'href': a.href, 'is_long': get_url_parts(a.href) }} )
+    
+    var a_counts = {}
+    as.forEach(function(a, i){
+        a_counts[a.href] = a_counts[a.href] || []
+        a_counts[a.href].push(i)
+    })
+    
+    var a_top_nodes = as.map(function(a, i){ 
+        return get_highest_singular_parent(i, as)
+    })
+'''
+
+js_to_spotcheck = '''
+    a_top_nodes.forEach(function(node){
+        node.setAttribute('style', 'border: 4px dotted blue !important;')
+    })
+'''
+
+
+def get_bounding_box_info(page):
+    bounding_boxes = page.evaluate('''    
+        function () {
+            var all_links = []
+            a_top_nodes.forEach(function(node){
+                var links = Array.from(node.querySelectorAll('a'))
+                if ((links.length == 0) & (node.nodeName === 'A')){
+                    links = [node]
+                }
+                
+                var seen_links = {};
+                links = links
+                    .map(function(a) {return {
+                        'href': a.href,
+                         'link_text' : get_text_of_node(a), 
+                        }
+                    } )
+                    .sort((a, b) => { return  b.link_text.length - a.link_text.length } )
+                    .filter(function(a){
+                        if (!(a.href in seen_links)) {
+                            seen_links[a.href] = true;
+                            return true
+                        }
+                        return false 
+                    })
+                    .forEach(function(a){
+                        var b = node.getBoundingClientRect() // get the bounding box around the entire defined node.
+                        a['x'] = b['x']
+                        a['y'] = b['y']
+                        a['width'] = b['width']
+                        a['height'] = b['height']
+                        a['all_text'] = get_text_of_node(node)
+                        a['css_attributes'] = getComputedStyle(node)
+                        all_links.push(a)
+                })
+            })
+            
+            seen_all_links = {}
+            return all_links.filter(function(a){
+                if (!([a.href, a.x, a.y] in seen_all_links)) {
+                    seen_all_links[[a.href, a.x, a.y]] = true;
+                    return true;
+                }
+                return false;
+            })
+        }
+    ''')
+
+    width = page.evaluate('''
+        Math.max(
+            document.documentElement["clientWidth"],
+            document.body["scrollWidth"],
+            document.documentElement["scrollWidth"],
+            document.body["offsetWidth"],
+            document.documentElement["offsetWidth"]
+        );
+    ''')
+
+    height = page.evaluate('''Math.max(
+        document.documentElement["clientHeight"],
+        document.body["scrollHeight"],
+        document.documentElement["scrollHeight"],
+        document.body["offsetHeight"],
+        document.documentElement["offsetHeight"]
+    );''')
+
+    return bounding_boxes, width, height
+
+
+def load_helper_scripts(page):
+    """Read and return Javascript code from a file. Convenience function."""
+    utils_script = utils.BIN_DIR / "js" / "psl.min.js"
+    with open(utils_script ) as f:
+        page.evaluate(f.read())
+
+    utils_script = utils.BIN_DIR / "js" / "utils.js"
+    with open(utils_script ) as f:
+        page.evaluate(f.read())
 
 
 @click.command()
@@ -33,43 +135,33 @@ def cli(handle: str, output_dir: str, timeout: str = "180"):
         context.close()
 
     # Write out the data
-    output_path = Path(output_dir) / f"{site['handle'].lower()}.hyperlinks.json"
+    output_path = Path(output_dir) / f"{site['handle'].lower()}.hyperlinks-with-bb.json"
     utils.write_json(link_list, output_path)
 
 
 @retry(tries=3, delay=5, backoff=2)
 def _get_links(context: BrowserContext, data: typing.Dict, timeout: int = 180):
-    print(f"🔗 Getting hyperlinks from {data['url']}")
+    print(f"Getting hyperlinks from {data['url']}")
     # Open a page
     page = context.new_page()
 
     # Go to the page
     page.goto(data["url"], timeout=timeout * 1000)
 
-    # Pull the html
-    html = page.content()
+    # load helper scripts into the page and get resources to run the rest of the scripts
+    load_helper_scripts(page)
 
-    # Parse out all the links
-    soup = BeautifulSoup(html, "html5lib")
-    link_list = soup.find_all("a")
+    # for each <a> on the page, get the upper-most child in the DOM that doesn't have any other links in the subtree
+    page.evaluate(get_link_divs_js)
 
-    # Parse out the data we want to keep
-    data_list = []
-    for link in link_list:
-        try:
-            d = dict(text=link.text, url=link["href"])
-        except KeyError:
-            # If no href, skip it
-            continue
-
-        # Add to big list
-        data_list.append(d)
+    # retrieve this data from the page object
+    bb, w, h = get_bounding_box_info(page)
 
     # Close the page
     page.close()
 
     # Return the result
-    return data_list
+    return {'bounding boxes': bb, 'page_width': w, 'page_height': h}
 
 
 if __name__ == "__main__":
